+++ robot2/input_data.h
@@ -6,6 +6,6 @@
 bool check_obstacle(float dist_obs);
 
 //restituisce il percoso leto dai linefollower
-int read_path(int sensorStateDX, int sensorStateSX);
+uint8_t read_path(uint8_t sensorStateDX, uint8_t sensorStateSX);
 
 #endif
